URO-84 URO-111 URO-113 URO-114 URO-115 URO-118
This commit structures the <Navigator /> component and uses placeholders to
stand in for components to be added later. Each component is described below.
It also adds more sample items for testing.

URO-84  <FilterContainer />
This component is completely determined by the search and sort props.

URO-111 <HeaderTabs />
This component must be aware of the location so that it may preserve previously
set query parameters. The category is received as a prop and determines the
state of this component.

URO-113 <NarrativeNewButton />

URO-114 <HeaderContainer />
Takes category search and sort props.

URO-115 <MainContainer />
Takes limit, items, narrative and view as props. If no items are passed then a
message is provided.

URO-118 <HeaderNavigationContainer />
This component is completely determined by the category prop.