URO-112 Create Refresh button
This introduces the navigator.fresh state which is a boolean representing that
the items held in navigator.items are up to date with user input. When they are
out of date navigator.fresh is false. The refresh button provides the user with
the ability to set this for any reason, such as when a narrative is shared by
another user.