URO-110 Create Sort Selection component
This commit adds the <SortSelect /> component which allows the user to specify
the order in which the list of narratives is arranged. None of this sorting
logic is present since the <NarrativeList /> component is meant to be handed a
sorted list of NarrativeDocs and will do no sorting of its own.

The path and search parameters are parsed to determine the category, which
narrative object is selected and the search and sort criteria as well as the
item limit. The search parameter will be used by the <SearchInput /> component
in a subsequent commit, and similarly the limit parameter will be used by the
<NarrativeList /> component.