Create a paramsSlice to manage search parameters.
This commit adds a paramsSlice to manage search parameters in the URL. Only
search parameters included in the slice are allowed and  URLs using others will
be redirected to URLs which contain only known search parameters. A log in the
console notes which parameters are ignored. This should prevent developers from
adding query parameters without adding them to the slice, or if they do notify
them that they are ignored. This also makes each query parameter global across
Europa so that the type of each parameter must be uniform and specified.