Add generatePathWithSearchParams to simplify links.
This commit adds the generatePathWithSearchParams function which takes a route
path specification and an object of parameters then fills in the path with them
and any extra are added as search parameters. This commit makes use of the
filtered parameters, so any extra which are not recognized will be ignored.
This replaces the previous keepParams function and derivatives.

This commit also siplifies the visibility logic for the NarrativeItemDropdown
component and changes the location of route path specifications related to Navigator.
In the Navigator module they are imported from the common module, but if they
are imported elsewhere within the app they should be imported from the
Navigator module.

Another change is to use destructuring assignment defaults when possible.

Finally, tests are added and updated for coverage and accuracy.