URO-143: Added searchapi2 using RTK Query.
This commit adds support for making requests to searchapi2 and caching their
results using RTK Query. The only supported request at present is searching for
a list of narratives as a part of Navigator. The <RefreshButton /> component
has been updated to invalidate these searches from the cache.

An important update is that the `id` field in JSONRPC requests has been
converted to a string type, which is mutually supported by JSONRPC 1.1 and
JSONRPC 2.0.