Add <NarrativeView /> component.
This commit adds the following components:

- <DefaultIcon />
- <MarkdownCellCard />
- <CellCard />
- <NarrativePreview />
- <NarrativeView />
- <NarrativeViewTabs />

The <CellCard /> component renders a card from a cell in a narrative. It
explicitly supports markdown, app, code and data cells and will do its best in
the face of ambiguity.

The <NarrativePreview /> component renders a list of up to 16 cell cards
associated with a workspace.

The <NarrativeView /> component provides a scaffold to hold the data and
preview narrative views, as well as placeholders for the control menu,
narrative metadata and version selection components.

The <NarrativeViewTabs /> component renders links for switching between data
and preview narrative views. The <DefaultIcon /> component provides a default
icon for <CellCard />s. The <MarkdownCellCard /> component renders a card from
markdown cell in a narrative.

Additionally, the <MainComponent /> in the Navigator module has been cleaned up which
included updating the <NarrativeList /> component props. There is a bug with
icons which will be addressed in the future. Enough CSS is included to roughly
align the layout to the current Navigator design.