URO-149: Connect data and preview views to API
This commit makes use of the getwsNarrative workspaceApi method to load
narrative data from the workspace for a given narrative.

This commit adds the useCells hook to load the cells from the workspaceApi
response into the state. Additionally, the list of the most recent workspace
objects returned by the searchApi is also loaded into the state. With the cells
and data objects in the state, the <NarrativeView /> component now shows the
current state of the workspace for the selected narrative.

Finally, NarrativeListDoc is deprecated and NarrativeDoc, the type representing
narrative objects returned from searchapi2, is preferred in all cases. The
NarrativeListDoc type lives on as NarrativeTestDoc to aid construction of the
NarrativeList.fixture, in particular, a list of NarrativeDocs for testing. The
factoryNarrativeDoc function takes a NarrativeTestDoc and lifts it to a
NarrativeDoc, minimizing the exposure of NarrativeTestDoc beyond the context of
tests. Ultimately, this type aids in reducing the amount of code required to
produce testing data.