remove material icons, use fa chevron icon instead [URO-208]
had followed the documentation which uses mui icons (and the docs have warnings and workarounds for fa icons,which may not be relevant here), but the fa icon does indeed work.  Also overrides the default orientation and rotation, which matches the other usage in the codebase which itself matches the kbase-ui usage which matches the Narrative which uses an augmented bootstrap 3 "accordion". Notable, though that BS5 Accordions work the same as MUI (BS 3 and 4 do not use icon at all.)

Also added the recommended aria- attributes.